Move icon templates to 'icons' directory
