+++ resources/ts/templates/icons/copyIcon.ts
