+++ resources/ts/templates/icons/searchIcon.ts
